% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdnet.R
\name{cleanDatalist}
\alias{cleanDatalist}
\title{Clean the list of datasets}
\usage{
cleanDatalist(datalist, geneset = NULL)
}
\arguments{
\item{datalist}{A list of datasets}

\item{geneset}{An optional list of row names to reduce the datasets by}
}
\value{
A list of datasets that have matching rows
}
\description{
Takes a list of datasets, finds the nodes that are common across all, and matches the rows so they are in the same format.
}

