% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdnet.R
\name{cmdist}
\alias{cmdist}
\title{Correlation matrix distance}
\usage{
cmdist(r1, r2, type = c("network", "node"), metric = c("angle", "cosine"))
}
\arguments{
\item{r1}{The first correlation matrix}

\item{r2}{The second correlation matrix}

\item{type}{whether to compute statistics at the network or node level}

\item{metric}{whether to use the angular distance or the cosine distance}
}
\value{
A number or vector of distances, depending on the type of computation specified.
}
\description{
Computes the correlation matrix distance between two correlation matrices.
}

