% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdnet.R
\name{cmdnet}
\alias{cmdnet}
\title{Correlation matrix distance across networks}
\usage{
cmdnet(datalist, genelist = NULL, mode = c("cmd", "node", "net"),
  corFUN = cor, xform = c("none", "fisher"), r0 = NULL,
  metric = c("angle", "cosine"))
}
\arguments{
\item{datalist}{A list of datasets}

\item{genelist}{An optional list of row names to reduce the datasets by}

\item{mode}{Which of 3 computations to perform. Defaults to all 3.}

\item{corFUN}{A function specifying which correlation measure to use. Defaults to "cor"}

\item{xform}{Whether to transform corrlations using Fisher's z transform. Defaults to "none".}

\item{r0}{An optional consensus matrix to compute deviations around. Defaults to NULL, and computes average matrix.}

\item{metric}{Whether to use angular distance or cosine distance. Defaults to "angle".}
}
\value{
A list of correlation matrix distance statistics
}
\description{
Computes correlation matrix distance statistics for a list of datasets.
}

